+++ acquire-core-libs/src/acquire-device-properties/device/props/storage.h
@@ -11,6 +11,34 @@ extern "C"
     struct DeviceManager;
     struct VideoFrame;
 
+    enum DimensionType
+    {
+        DimensionType_Space = 0,
+        DimensionType_Channel,
+        DimensionType_Time,
+        DimensionType_Other,
+        DimensionTypeCount
+    };
+
+    struct StorageDimension
+    {
+        // the name of the dimension as it appears in the metadata, e.g.,
+        // "x", "y", "z", "c", "t"
+        struct String name;
+
+        // the type of dimension, e.g., spatial, channel, time
+        enum DimensionType kind;
+
+        // the expected size of the full output array along this dimension
+        uint32_t array_size_px;
+
+        // the size of a chunk along this dimension
+        uint32_t chunk_size_px;
+
+        // the number of chunks in a shard along this dimension
+        uint32_t shard_size_chunks;
+    };
+
     /// Properties for a storage driver.
     struct StorageProperties
     {
@@ -19,17 +47,23 @@ extern "C"
         uint32_t first_frame_id;
         struct PixelScale pixel_scale_um;
 
-        /// Dimensions of chunks, in pixels.
-        struct storage_properties_chunking_s
+        /// Dimensions of the output array, with array extents, chunk sizes, and
+        /// shard sizes. The first dimension is the fastest varying dimension.
+        /// The last dimension is the append dimension.
+        struct storage_properties_dimensions_s
         {
-            uint32_t width, height, planes;
-        } chunk_dims_px;
+            // The dimensions of the output array.
+            struct StorageDimension* data;
 
-        /// Dimensions of shards, in chunks.
-        struct storage_properties_sharding_s
-        {
-            uint32_t width, height, planes;
-        } shard_dims_chunks;
+            // The number of dimensions in the output array.
+            size_t size;
+
+            // Allocate storage for the dimensions.
+            void (*init)(struct StorageDimension**, size_t);
+
+            // Free storage for the dimensions.
+            void (*destroy)(struct StorageDimension*);
+        } acquisition_dimensions;
 
         /// Enable multiscale storage if true.
         uint8_t enable_multiscale;
@@ -37,28 +71,9 @@ extern "C"
 
     struct StoragePropertyMetadata
     {
-        /// Metadata for chunking.
-        /// Indicates whether chunking is supported, and if so, bounds on what
-        /// the dimensions (in px) of the chunks are.
-        struct storage_property_metadata_chunking_s
-        {
-            uint8_t is_supported;
-            struct Property width, height, planes;
-        } chunk_dims_px;
-
-        /// Metadata for sharding.
-        /// Indicates whether sharding is supported, and if so, bounds on what
-        /// the dimensions (in chunks) of the shards are.
-        struct storage_property_metadata_sharding_s
-        {
-            uint8_t is_supported;
-            struct Property width, height, planes;
-        } shard_dims_chunks;
-
-        struct storage_property_metadata_multiscale_s
-        {
-            uint8_t is_supported;
-        } multiscale;
+        uint8_t chunking_is_supported;
+        uint8_t sharding_is_supported;
+        uint8_t multiscale_is_supported;
     };
 
     /// Initializes StorageProperties, allocating string storage on the heap
@@ -114,34 +129,6 @@ extern "C"
                                                  const char* metadata,
                                                  size_t bytes_of_metadata);
 
-    /// @brief Set chunking properties for `out`.
-    /// Convenience function to set chunking properties in a single call.
-    /// @returns 1 on success, otherwise 0
-    /// @param[in, out] out The storage properties to change.
-    /// @param[in] chunk_width The width, in px, of a chunk.
-    /// @param[in] chunk_height The height, in px, of a chunk.
-    /// @param[in] chunk_planes The number of @p chunk_width x @p chunk_height
-    ///            planes in a single chunk.
-    int storage_properties_set_chunking_props(struct StorageProperties* out,
-                                              uint32_t chunk_width,
-                                              uint32_t chunk_height,
-                                              uint32_t chunk_planes);
-
-    /// @brief Set sharding properties for `out`.
-    /// Convenience function to set sharding properties in a single call.
-    /// @returns 1 on success, otherwise 0
-    /// @param[in, out] out The storage properties to change.
-    /// @param[in] shard_width The number of chunks in a shard along the x
-    ///            dimension.
-    /// @param[in] shard_height The number of chunks in a shard along the y
-    ///            dimension.
-    /// @param[in] shard_planes The number of chunks in a shard along the append
-    ///            dimension.
-    int storage_properties_set_sharding_props(struct StorageProperties* out,
-                                              uint32_t shard_width,
-                                              uint32_t shard_height,
-                                              uint32_t shard_planes);
-
     /// @brief Set multiscale properties for `out`.
     /// Convenience function to enable multiscale.
     /// @returns 1 on success, otherwise 0
@@ -152,6 +139,50 @@ extern "C"
 
     /// Free's allocated string storage.
     void storage_properties_destroy(struct StorageProperties* self);
+
+    /// @brief Initialize the Dimension struct in `out`.
+    /// @param[out] out The Dimension struct to initialize.
+    /// @param[in] name The name of the dimension.
+    /// @param[in] bytes_of_name The number of bytes in the name buffer.
+    ///                          Should include the terminating NULL.
+    /// @param[in] kind The type of dimension.
+    /// @param[in] array_size_px The size of the array along this dimension.
+    /// @param[in] chunk_size_px The size of a chunk along this dimension.
+    /// @param[in] shard_size_chunks The number of chunks in a shard along this
+    ///                              dimension.
+    /// @returns 1 on success, otherwise 0
+    int storage_dimension_init(struct StorageDimension* out,
+                               const char* name,
+                               size_t bytes_of_name,
+                               enum DimensionType kind,
+                               uint32_t array_size_px,
+                               uint32_t chunk_size_px,
+                               uint32_t shard_size_chunks);
+
+    /// @brief Copy the Dimension struct in `src` to `dst`.
+    /// @param[out] dst The Dimension struct to copy to.
+    /// @param[in] src The Dimension struct to copy from.
+    /// @returns 1 on success, otherwise 0
+    int storage_dimension_copy(struct StorageDimension* dst,
+                               const struct StorageDimension* src);
+
+    /// @brief Destroy the Dimension struct in `self`.
+    /// @param[out] self The Dimension struct to destroy.
+    void storage_dimension_destroy(struct StorageDimension* self);
+
+    /// @brief Initialize the acquisition_dimensions array in `self`.
+    /// @param[out] self The StorageProperties struct containing the array to
+    ///                  initialize.
+    /// @param[in] size The number of dimensions to allocate.
+    /// @returns 1 on success, otherwise 0
+    int storage_properties_dimensions_init(struct StorageProperties* self,
+                                           size_t size);
+
+    /// @brief Free the acquisition_dimensions array in `self`.
+    /// @param[out] self The StorageProperties struct containing the array to
+    ///                  destroy.
+    /// @returns 1 on success, otherwise 0
+    int storage_properties_dimensions_destroy(struct StorageProperties* self);
 
 #ifdef __cplusplus
 }